Merge old document into docstring
I ran `lychee` on the repository, and it found that a linked Google
Doc[^1] was inaccessible. I was able to access it, but I did not want
our repository to rely on a private document in the Google Drive of a
former student on the project.

It was quite short, so I used `pandoc` to convert the document to
Markdown, which I then manually cleaned up and added to the relevant
module's docstring.

[^1]: https://docs.google.com/document/d/16RODa6KDX7vNNooBdciI4NqSVN31lToto3MLTNcEHk0/edit