Change to newer JSON suggestion format
The first JSON format was designed to allow easy filtering by the
intended recipient, but it assumed that a player would only receive
advice about themselves. It was unable to handle move advice about other
players, which we intend to support, so I needed to redesign the format
to allow it.

The new format has an explicit `recipient` field to avoid accidentally
intercepting any suggestions intended for another player. In addition,
it can handle other top-level metadata that we want to store, such as
the advisor that sent the suggestion.