Prohibit suggestions when bot does not declare support
All bots should declare what suggestion types they support in
`BaselineBot.suggestion_type`, and that value is needed so the UI
displays the sent suggestions.

We should raise an exception to prevent a bot from accidentally sending
suggestions that the user would never see when the suggestion support is
not explicitly declared.