Decrease delay between running agents from 4 s to 1 s
I realized that waiting 4 s between each agent meant waiting 24 s
between starting the first and last agent. Reducing it to 1 s reduces
that wait time down to 6 s. I tried removing the delay entirely, but
that appeared to cause some sort of contention problem.

I decided to make the delay configurable because the higher specs of my
machine compared to when I wrote the code is likely what allowed me to
decrease it. Anyone else who runs the script might need to manually set
it back to 4 s to prevent problems.