Improve typing for `suggest_orders_probabilities()`
The method's interface was undocumented and relied on passing in the
output of `baseline-models` directly. Now that I am adding support for
order probability advice to Cicero, the interface needs to actually be
documented so I can construct data in the proper format.

The change was mostly straightforward, but the best way to properly
document a `dict`'s structure is the use of `TypedDict`, which was
introduced in Python 3.8. Because Cicero is stuck on Python 3.7, I
needed to add `typing_extensions` as a direct dependency. This doesn't
really change anything because it was an indirect requirement, anyway.