Download FAAF model from Hugging Face Hub
Instead of hosting the model in Google Drive, we should use the model
directly from the Hugging Face Hub. This is more reliable long-term
storage, and it makes downloading the model automatic.

As part of this commit, I renamed some variables to account for the fact
that the model is no longer loaded from a path.